backend/requirements.txt
django==4.2
djangorestframework==3.14
django-cors-headers==4.3
pymongo==4.6
python-dotenv==1.0
google-auth==2.27
google-auth-oauthlib==1.2
google-api-python-client==2.118
google-generativeai==0.4
requests==2.31
